Finish setting up admin menu, register scripts, display hello world.
