+++ plugins/woocommerce-docs/client/index.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 const App = () => {
-	return <h1>Hello World</h1>;
+	return <h1>WooCommerce Docs Administration</h1>;
 };
 
 // render the app with react
