--- plugins/woocommerce-docs/includes/admin/class-docs-menu.php
@@ -1,44 +1,0 @@
-<?php
-/**
- * Docs_Menu class file.
- *
- * @package  WooCommerce_Docs
- */
-
-/**
- * A class to set up menus for this WordPress plugin
- */
-class Docs_Menu {
-
-	/**
-	 * Define the constructor
-	 */
-	public function __construct() {
-		add_action( 'admin_menu', array( $this, 'add_admin_menu' ) );
-	}
-
-	/**
-	 * Define the add_admin_menu function
-	 */
-	public function add_admin_menu() {
-		// Add a top-level menu item to the admin menu.
-		add_menu_page(
-			'WooCommerce Docs',
-			'WooCommerce Docs',
-			'manage_options',
-			'woocommerce-docs',
-			array( $this, 'render_admin_page' ),
-			'dashicons-media-document',
-			6
-		);
-	}
-
-	/**
-	 * Render admin page
-	 */
-	public function render_admin_page() {
-		// Include the admin page template.
-		include_once plugin_dir_path( __FILE__ ) . 'includes/views/admin.php';
-	}
-}
-
