+++ plugins/woocommerce-docs/tsconfig.json
@@ -1,0 +1,27 @@
+{
+	"compilerOptions": {
+		// Target latest version of ECMAScript.
+		"target": "esnext",
+
+		// Make sure we're not pulling types from external projects.
+		"typeRoots": [ "./typings", "./node_modules/@types" ],
+
+		// Search under node_modules for non-relative imports.
+		"moduleResolution": "node",
+
+		"jsx": "react-jsx",
+
+		// Temporary for resolving the test data json, remove when moving to using the API
+		"resolveJsonModule": true,
+
+		// Enable strictest settings like strictNullChecks & noImplicitAny.
+		"strict": true,
+
+		// Import non-ES modules as default imports.
+		"esModuleInterop": true,
+
+		// Skip type checking of declaration files because some libraries define two copies of the same type in an inconsistent way
+		"skipLibCheck": true,
+		"module": "esnext"
+	}
+}
