+++ plugins/woocommerce-docs/woocommerce-docs.php
@@ -9,4 +9,7 @@
  **/
 require __DIR__ . '/vendor/autoload.php';
 
-new \WooCommerce_Docs\Docs_Menu();
+define( 'WOOCOMMERCE_DOCS_ROOT_URL', plugin_dir_url( __FILE__ ) );
+const WOOCOMMERCE_DOCS_PLUGIN_PATH = __DIR__;
+
+new \WooCommerce_Docs\Admin\Docs_Menu();
